+++ src/pydocformatter/rules/definitions/PCF/__init__.py
@@ -1,1 +1,3 @@
-"""PCF rule definitions."""
+from pydocformatter.rules.base import RuleLinterMetadata
+
+LINTER = RuleLinterMetadata(prefix="PCF", name="pydocformatter comment formatting", url="https://github.com/pallgeuer/pydocformatter")