+++ src/pydocformatter/rules/definitions/PDF/__init__.py
@@ -1,1 +1,3 @@
-"""PDF rule definitions."""
+from pydocformatter.rules.base import RuleLinterMetadata
+
+LINTER = RuleLinterMetadata(prefix="PDF", name="pydocformatter docstring formatting", url="https://github.com/pallgeuer/pydocformatter")